+++ CHANGELOG.md
@@ -50,6 +50,9 @@
                                   , "ReturnType" )                   ; not need
   ```
 
+- feat: ([#18](https://github.com/CoffeeChaton/vscode-autohotkey-NekoHelp/issues/18)) support InlayHints Provider
+  > draft <https://www.youtube.com/watch?v=qvPI2tf8OLc>
+
 ## v0.0.51(2023-12-03)
 
 <!-- pnpm 8.10.2 -->